Moved ParserTest and QueryToStringTest to their respective package in Graql
